Patches made at the ChemMaster 3000 now function as a bandaid

+++ code/modules/reagents/reagent_containers/patch.dm
@@ -36,12 +36,25 @@
 			if(!H.can_inject(user, FALSE, L.zone_sel.selecting, pierce_material))
 				to_chat(user, "<span class='notice'>\The [src] can't be applied through such a thick material!</span>")
 				return
+			
+			if(affecting.open)// you cant place Bandaids on open surgeries, why chemical patches.
+				to_chat(user, "<span class='notice'>The [affecting.name] is cut open, you'll need more than a bandage!</span>")
+				return
 
 			to_chat(H, "<span class='notice'>\The [src] is placed on your [affecting].</span>")
 			M.drop_from_inventory(src) //icon update
 			if(reagents.total_volume)
 				reagents.trans_to_mob(M, reagents.total_volume, CHEM_TOUCH)
 			qdel(src)
+
+			for(var/datum/wound/W in affecting.wounds)
+				if (W.internal)//ignore internal wounds and check the remaining
+					continue
+				if(W.bandaged)//already bandaged wounds dont need to be bandaged again
+					continue
+				W.bandage()
+				break;//dont bandage more than one wound, its only one patch you can have in your stack
+			affecting.update_damages()
 			return 1
 
 	else if(istype(M, /mob/living/carbon/human))
@@ -56,6 +69,9 @@
 
 		if(!H.can_inject(user, FALSE, L.zone_sel.selecting, pierce_material))
 			to_chat(user, "<span class='notice'>\The [src] can't be applied through such a thick material!</span>")
+			return
+		if(affecting.open)// you cant place Bandaids on open surgeries, why chemical patches.
+			to_chat(user, "<span class='notice'>The [affecting.name] is cut open, you'll need more than a bandage!</span>")
 			return
 
 		user.visible_message("<span class='warning'>[user] attempts to place \the [src] onto [H]`s [affecting].</span>")
@@ -76,6 +92,15 @@
 		if(reagents.total_volume)
 			reagents.trans_to_mob(M, reagents.total_volume, CHEM_TOUCH)
 		qdel(src)
+
+		for(var/datum/wound/W in affecting.wounds)
+			if (W.internal)//ignore internal wounds and check the remaining
+				continue
+			if(W.bandaged)//already bandaged wounds dont need to be bandaged again
+				continue
+			W.bandage()
+			break;//dont bandage more than one wound, its only one patch you can have in your stack
+		affecting.update_damages()
 
 		return 1
